Export Element Labels
Since it was related to translation decided to add the solution to this problem here. This is a referance to which problem im referencing since it didnt have linked issue number:

Problem: Currently, to "get" all content of a grid set one has to go to 'translate grid', choose the 'show all grids' from the first dropdown and then use the 'copy column' option to get the words.